Add GitLab CI configuration file
This commit adds a new `.gitlab-ci.yaml` file to the repository. The file includes configuration for two stages: `test` and `publish`. In the `test` stage, it runs the command `npm run test`. In the `publish` stage, it sets up authentication for publishing to the GitLab Package Registry and then runs `npm publish`. This configuration will only be triggered on the `main` branch.